+++ src/utils/shuffleArray.ts
@@ -1,0 +1,11 @@
+/**
+ * Fisher-Yates shuffle. Returns a new array; the input is not mutated.
+ */
+export function shuffleArray<T>(items: T[]): T[] {
+  const arr = [...items];
+  for (let i = arr.length - 1; i > 0; i--) {
+    const j = Math.floor(Math.random() * (i + 1));
+    [arr[i], arr[j]] = [arr[j], arr[i]];
+  }
+  return arr;
+}
