+++ src/utils/songToTrack.ts
@@ -1,0 +1,28 @@
+import type { SubsonicSong } from '../api/subsonic';
+import type { Track } from '../store/playerStore';
+
+export function songToTrack(song: SubsonicSong): Track {
+  return {
+    id: song.id,
+    title: song.title,
+    artist: song.artist,
+    album: song.album,
+    albumId: song.albumId,
+    artistId: song.artistId,
+    duration: song.duration,
+    coverArt: song.coverArt,
+    track: song.track,
+    year: song.year,
+    bitRate: song.bitRate,
+    suffix: song.suffix,
+    userRating: song.userRating,
+    replayGainTrackDb: song.replayGain?.trackGain,
+    replayGainAlbumDb: song.replayGain?.albumGain,
+    replayGainPeak: song.replayGain?.trackPeak,
+    starred: song.starred,
+    genre: song.genre,
+    samplingRate: song.samplingRate,
+    bitDepth: song.bitDepth,
+    size: song.size,
+  };
+}
